change round procedure code
Changed the code from regularly checking the round state and executing certain functions depending on the variable values to waiting for the asynchronous function to finish using async and await
